[w32handle] Move them to metadata
w32handle are purely for managed support, and they shouldn't be used in the runtime.
